update : package structure
