+++ Mobile Development/RelasiaHelperApp/app/src/main/java/com/c22ps099/relasiahelperapp/ui/login/LoginFragment.kt
@@ -1,4 +1,4 @@
-package com.c22ps099.relasiahelperapp.view
+package com.c22ps099.relasiahelperapp.ui.login
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
