+++ Mobile Development/RelasiaHelperApp/app/src/main/java/com/c22ps099/relasiahelperapp/ui/main/BaseActivity.kt
@@ -1,7 +1,8 @@
-package com.c22ps099.relasiahelperapp.view
+package com.c22ps099.relasiahelperapp.ui.main
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.c22ps099.relasiahelperapp.R
 
 class BaseActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
