+++ Mobile Development/RelasiaHelperApp/app/src/main/java/com/c22ps099/relasiahelperapp/ui/missions/MissionsFragment.kt
@@ -1,4 +1,4 @@
-package com.c22ps099.relasiahelperapp.view
+package com.c22ps099.relasiahelperapp.ui.missions
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -14,10 +14,10 @@ private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
- * Use the [DetailMissionFragment.newInstance] factory method to
+ * Use the [MissionsFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
-class DetailMissionFragment : Fragment() {
+class MissionsFragment : Fragment() {
     // TODO: Rename and change types of parameters
     private var param1: String? = null
     private var param2: String? = null
@@ -35,7 +35,7 @@ class DetailMissionFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_detail_mission, container, false)
+        return inflater.inflate(R.layout.fragment_missions, container, false)
     }
 
     companion object {
@@ -45,12 +45,12 @@ class DetailMissionFragment : Fragment() {
          *
          * @param param1 Parameter 1.
          * @param param2 Parameter 2.
-         * @return A new instance of fragment DetailMissionFragment.
+         * @return A new instance of fragment MissionsFragment.
          */
         // TODO: Rename and change types and number of parameters
         @JvmStatic
         fun newInstance(param1: String, param2: String) =
-            DetailMissionFragment().apply {
+            MissionsFragment().apply {
                 arguments = Bundle().apply {
                     putString(ARG_PARAM1, param1)
                     putString(ARG_PARAM2, param2)
