+++ Mobile Development/RelasiaHelperApp/app/src/main/res/layout/activity_base.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.BaseActivity">
+    tools:context=".ui.main.BaseActivity">
 
     <View
         android:layout_width="match_parent"
