+++ Mobile Development/RelasiaHelperApp/app/src/main/res/layout/item_mission.xml
@@ -18,9 +18,9 @@
         <ImageView
             android:id="@+id/iv_mission_photo"
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"
+            android:layout_height="240dp"
             android:layout_margin="24dp"
-            android:adjustViewBounds="true"
+            android:scaleType="centerCrop"
             android:contentDescription="@string/mission_photo"
             card_view:layout_constraintEnd_toEndOf="parent"
             card_view:layout_constraintStart_toStartOf="parent"
